# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Toolstead Platform API
  version: 0.4.0
  description: Workspace, module entitlement, and CRM foundation for Toolstead.
servers:
  - url: /api/v1
security:
  - accessCookie: []
paths:
...
  /work-queue:
    get:
      operationId: listWorkQueue
      parameters:
        - name: scope
          in: query
          schema:
            type: string
            enum: [attention, today, all]
            default: attention
        - name: q
          in: query
          schema:
            type: string
            maxLength: 120
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        "200":
          description: Workspace work items.
          content:
            application/json:
              schema:
                type: object
                required: [scope, count, items]
                properties:
                  scope:
                    type: string
                  count:
                    type: integer
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/WorkItem"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /contacts/{contactId}/timeline:
    get:
      operationId: getContactTimeline
      parameters:
        - $ref: "#/components/parameters/ContactId"
      responses:
        "200":
          description: Contact timeline.
        "400":
          $ref: "#/components/responses/BadRequest"
...
  /contacts/{contactId}/messages:
    post:
      operationId: queueContactMessage
      parameters:
        - $ref: "#/components/parameters/ContactId"
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
            minLength: 8
            maxLength: 128
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [channel, body]
              properties:
                channel:
                  type: string
                  enum: [email, sms]
                subject:
                  type: string
                  maxLength: 200
                body:
                  type: string
                  minLength: 1
                  maxLength: 10000
      responses:
        "202":
          description: Message accepted into the outbox or the idempotent prior result returned.
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: Consent or destination conflict.
  /contacts/{contactId}/consents:
    post:
      operationId: recordContactConsent
      parameters:
        - $ref: "#/components/parameters/ContactId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConsentInput"
      responses:
        "201":
          description: Consent evidence recorded.
        "403":
          $ref: "#/components/responses/Forbidden"
  /contacts/{contactId}/appointments:
    post:
      operationId: createContactAppointment
      parameters:
        - $ref: "#/components/parameters/ContactId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentInput"
      responses:
        "201":
          description: Appointment created.
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: Appointment overlaps an existing requested or confirmed appointment.
components:
  securitySchemes:
    accessCookie:
      type: apiKey
      in: cookie
      name: toolstead_access
    refreshCookie:
      type: apiKey
      in: cookie
      name: toolstead_refresh
  parameters:
    ContactId:
      name: contactId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    BadRequest:
      description: Request validation failed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Authentication is required or invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Role or module entitlement denied.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    WorkItem:
      type: object
      required:
        [id, contactId, name, initials, type, source, age, urgency, nextStep, action, actionType, timeline]
      properties:
        id:
          type: string
          format: uuid
        contactId:
          type: string
          format: uuid
        name:
          type: string
        initials:
          type: string
        type:
          type: string
        source:
          type: string
        createdAt:
          type: string
          format: date-time
        age:
          type: string
        urgency:
          type: string
        urgencyTone:
          type: string
        nextStep:
          type: string
        note:
          type: [string, "null"]
        action:
          type: string
        actionType:
          type: string
        phone:
          type: [string, "null"]
        email:
          type: [string, "null"]
        consent:
          type: string
        summary:
          type: string
        nextAction:
          type: string
        timeline:
          type: array
          items:
            type: object
...
    ConsentInput:
      type: object
      additionalProperties: false
      required: [channels]
      properties:
        channels:
          type: array
          minItems: 1
          maxItems: 4
          description: Each channel value may appear only once per request.
          items:
            type: object
            additionalProperties: false
            required: [channel, status]
            properties:
              channel:
                type: string
                enum: [email, sms, voice, facebook]
              status:
                type: string
                enum: [granted, revoked, unknown]
              evidence:
                type: string
                maxLength: 1000
    AppointmentInput:
      type: object
      additionalProperties: false
      required: [title, startsAt, endsAt, timezone]
      properties:
        title:
          type: string
          minLength: 2
          maxLength: 180
        startsAt:
          type: string
          format: date-time
        endsAt:
          type: string
          format: date-time
        timezone:
          type: string
          minLength: 3
          maxLength: 80
        notes:
          type: string
          maxLength: 2000
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code:
              type: string
            message:
              type: string
            requestId:
              type: string